Minor fixes to comp_bdpatch and some other things. (#600)

+++ libcomp/src/ReadOnlyPacket.cpp
@@ -421,7 +421,7 @@ uint16_t ReadOnlyPacket::PeekU16Little()
 int16_t ReadOnlyPacket::PeekS16Little()
 {
     // Return the value converted from little endian byte order.
-    return (int16_t)le16toh(PeekS16());
+    return (int16_t)le16toh((uint16_t)PeekS16());
 }
 
 uint32_t ReadOnlyPacket::PeekU32()
@@ -479,7 +479,7 @@ uint32_t ReadOnlyPacket::PeekU32Little()
 int32_t ReadOnlyPacket::PeekS32Little()
 {
     // Return the value converted from little endian byte order.
-    return (int32_t)le32toh(PeekS32());
+    return (int32_t)le32toh((uint32_t)PeekS32());
 }
 
 uint8_t ReadOnlyPacket::ReadU8()
@@ -579,7 +579,7 @@ int16_t ReadOnlyPacket::ReadS16Big()
 int16_t ReadOnlyPacket::ReadS16Little()
 {
     // Return the value converted from little endian byte order.
-    return (int16_t)le16toh(ReadS16());
+    return (int16_t)le16toh((uint16_t)ReadS16());
 }
 
 uint32_t ReadOnlyPacket::ReadU32()
@@ -641,7 +641,7 @@ int32_t ReadOnlyPacket::ReadS32Big()
 int32_t ReadOnlyPacket::ReadS32Little()
 {
     // Return the value converted from little endian byte order.
-    return (int32_t)le32toh(ReadS32());
+    return (int32_t)le32toh((uint32_t)ReadS32());
 }
 
 uint64_t ReadOnlyPacket::ReadU64()
@@ -703,7 +703,7 @@ int64_t ReadOnlyPacket::ReadS64Big()
 int64_t ReadOnlyPacket::ReadS64Little()
 {
     // Return the value converted from little endian byte order.
-    return (int64_t)le64toh(ReadS64());
+    return (int64_t)le64toh((uint64_t)ReadS64());
 }
 
 float ReadOnlyPacket::ReadFloat()
